Refactor discovery.py into a package and name eval runs
Split the 1012-line single file into a proper package following the
evaluation/ pattern (flat modules, no underscore prefixes). Wrap the
two mlflow.genai.evaluate() calls with mlflow.start_run(run_name=...)
so triage and validation runs are identifiable in the UI.
