+++ mlflow/genai/discovery/__init__.py
@@ -1,0 +1,4 @@
+from mlflow.genai.discovery.entities import DiscoverIssuesResult, Issue
+from mlflow.genai.discovery.pipeline import discover_issues
+
+__all__ = ["discover_issues", "DiscoverIssuesResult", "Issue"]
